refactor: Move away custom hooks
